Move Mesh Generation off Main thread. Not a working solution.
Trying to smooth out the frames by not letting mesh generation block the
thread. However, Meshs can't be instantiated off the main thread so I
still need to create a container that will hold the arrays generated in
the background thread and feed it to the Main thread for mesh
generation.